some functions in to do controller modification

+++ ToDoList/Repositories/ToDoItemRepos/ToDoItemDapperRepo.cs
@@ -21,40 +21,42 @@ namespace ToDoList.Repositories.ToDoItemRepos
         public async Task<ToDoItemtEntity> CreateToDoItemAsync(ToDoItemtEntity toDodb)
         {
             string sql = "Insert into Lists (ItemID,ItemName,IsFinished,CreatedDate,UserID,EndedDate) VALUES (@ItemID,@ItemName,@IsFinished,@CreatedDate,@UserID,@EndedDate)";
-            var sql1 = "SELECT * FROM Lists WHERE ItemID = @Id";
-            var sql2 = "SELECT * FROM Lists WHERE UserID = @Uid";
+            string sql1 = "SELECT ItemID FROM Lists WHERE ItemID = @Id";
+            string sql2 = "SELECT UserID FROM Lists WHERE UserID = @Uid";
             //string sql = @"
             //Insert into Lists (ItemID,ItemName,IsFinished,CreatedDate,UserID,EndedDate) VALUES (@ItemID,@ItemName,@IsFinished,@CreatedDate,@UserID,@EndedDate);
             //SELECT /*CAST( SCOPE_IDENTITY() as int*/ ItemID from Lists/*)*/";
 
             using IDbConnection connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection"));
             await connection.ExecuteAsync(sql, toDodb);
-            var result = await connection.QueryFirstOrDefaultAsync<int>(sql1, new { Id = toDodb.ItemID/*, Uid = toDodb.UserID*/ });
-            var result1 = await connection.QueryFirstOrDefaultAsync<int>(sql2, new { Uid = toDodb.UserID /*, Id = toDodb.ItemID*/ });
+            var result = await connection.QueryFirstOrDefaultAsync<int>(sql1, new { Id = toDodb.ItemID });
+            var result1 = await connection.QueryFirstOrDefaultAsync<int>(sql2, new { Uid = toDodb.UserID });
             var output = await GetToDoByIdAsync(result,result1);
             return output;
 
         }
 
-        public async Task DeleteToDoByIdAsync(int id)
+        public async Task DeleteToDoByIdAsync(int id , int uid)
         {
-            var sql = "DELETE FROM Lists WHERE ItemID = @Id";
+            var sql = "DELETE FROM Lists WHERE ItemID = @Id and UserID = @Uid";
             using IDbConnection connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection"));
 
-            await connection.ExecuteAsync(sql, new { Id = id });
+            await connection.ExecuteAsync(sql, new { Id = id, Uid = uid });
 
         }
 
         public async Task<ToDoItemtEntity> EditToDoByIdAsync(ToDoItemtEntity toDodb)
         {
-            string sql = @"
-            UPDATE Lists SET ItemName = @ItemName, IsFinished = @IsFinished, CreatedDate = @CreatedDate, UserID = @UserID, EndedDate = @EndedDate where ItemID = @ItemID;
-            SELECT CAST(SCOPE_IDENTITY() as int)";
-            //var sql = "UPDATE Lists SET ItemName = @ItemName, IsFinished = @IsFinished, CreatedDate = @CreatedDate, UserID = @UserID, EndedDate = @EndedDate where ItemID = @ItemID";
+            
+            var sql = "UPDATE Lists SET ItemName = @ItemName, IsFinished = @IsFinished, CreatedDate = @CreatedDate, EndedDate = @EndedDate where ItemID = @ItemID and UserID = @UserID";
+            string sql1 = "SELECT ItemID FROM Lists WHERE ItemID = @Id";
+            string sql2 = "SELECT UserID FROM Lists WHERE UserID = @Uid";
             using IDbConnection connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection"));
 
             await connection.ExecuteAsync(sql, toDodb);
-            var output = await GetToDoByIdAsync(toDodb.ItemID , toDodb.UserID);
+            var result = await connection.QueryFirstOrDefaultAsync<int>(sql1, new { Id = toDodb.ItemID });
+            var result1 = await connection.QueryFirstOrDefaultAsync<int>(sql2, new { Uid = toDodb.UserID });
+            var output = await GetToDoByIdAsync(result, result1);
             return output;
         }
